CAmbio en la Bd

+++ src/main/resources/schema.sql
@@ -146,7 +146,7 @@ idActividad INTEGER,
 
 IdSocio INTEGER, 
 
-posicion NUMERIC 
+posicion INTEGER 
 
 );
